requirement.txt
flask
selenium
alembic
psycopg2
bcrypt
aoiklivereload
python-dotenv
sanic_cors
uvicorn
datetime
requests
jwt
